requirements.txt
numpy==1.19.5
keras
tensorflow-cpu
h5py
pillow
flask
gunicorn
flask_cors